+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 1.9.1 (not released yet)
 -----
 * aded ability to turn off annotation scanning
+* added ability to skip tasks during automatic builds
 
 1.9.0
 -----